Remove the javadoc jar, source jar is enough
After updating gradle we have some problems generating the javadoc jar.
It keeps complaining about missing packages. This is because remock does
not depend on junit, and I'd like it to stay that way!

(1) At this time we do not host the javadoc anywhere.
(2) And IDEs will anyway prefer the source jar.